Create LODMeshGenerator.py
This is an add-on that can be installed in Blender to easily create four levels of detail from a given mesh, which can be a time consuming process.  It will generate four FBX files that will be ready to import into Unity 3D and serve as different level of details for an object.